+++ dbt/macros/core/current_record_flag.sql
@@ -1,0 +1,3 @@
+{% macro current_record_flag(valid_to_expression) -%}
+    iff({{ valid_to_expression }} is null, true, false)
+{%- endmacro %}