+++ dbt/macros/core/reconciliation_status.sql
@@ -1,0 +1,6 @@
+{% macro reconciliation_status(key_expression) -%}
+    case
+        when {{ key_expression }} is null then 'UNMATCHED'
+        else 'MATCHED'
+    end
+{%- endmacro %}